Add first gtest case for reading ASDF files
